requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pillow==12.0.0
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
shiboken6==6.10.1
tkinterdnd2==0.4.3
